// src/app/ui/models/factory.model.ts
export interface factory{
    id?: number;
    providerId?: number;
    name?: string;
    phone?: string;
    contactName?: string;
    dept?: string;
    city?:string;
    zone?:string;
    address?: string;
    email?: string;
    latitude?: number;
    length?: number;
    haveSoto13System?: boolean;
    workTimes? : string;
    state?:string;
}